+++ tempodb/encoding/vparquet/block_search.go
@@ -58,6 +58,7 @@ func (b *backendBlock) openForSearch(ctx context.Context, opts common.SearchOpti
 	b.openMtx.Lock()
 	defer b.openMtx.Unlock()
 
+	// TODO: ctx is also cached when we cache backendReaderAt, not ideal but leaving it as is for now
 	backendReaderAt := NewBackendReaderAt(ctx, b.r, DataFileName, b.meta.BlockID, b.meta.TenantID)
 
 	// no searches currently require bloom filters or the page index. so just add them statically
@@ -81,6 +82,7 @@ func (b *backendBlock) openForSearch(ctx context.Context, opts common.SearchOpti
 		}
 	}
 
+	// this is where we decide on a reader to use when opening a block for search...
 	// optimized reader
 	readerAt = newParquetOptimizedReaderAt(readerAt, int64(b.meta.Size), b.meta.FooterSize)
 
@@ -284,6 +286,7 @@ func searchParquetFile(ctx context.Context, pf *parquet.File, req *tempopb.Searc
 	return &tempopb.SearchResponse{
 		Traces:  results,
 		Metrics: &tempopb.SearchMetrics{},
+		// FIXME: SearchMetrics is empty here????
 	}, nil
 }